Renamed the media shared library
