+++ nx-secretarium-website/jest.config.js
@@ -4,5 +4,6 @@ module.exports = {
     '<rootDir>/libs/store/ui-shared',
     '<rootDir>/libs/store/media',
     '<rootDir>/libs/store/ui-data',
+    '<rootDir>/libs/store/ui-media',
   ],
 };
