--- nx-secretarium-website/libs/store/media/README.md
@@ -1,7 +1,0 @@
-# store-media
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test store-media` to execute the unit tests via [Jest](https://jestjs.io).
