--- nx-secretarium-website/libs/store/media/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/store-media';
