--- nx-secretarium-website/libs/store/media/src/lib/store-media.spec.ts
@@ -1,7 +1,0 @@
-import { storeMedia } from './store-media';
-
-describe('storeMedia', () => {
-  it('should work', () => {
-    expect(storeMedia()).toEqual('store-media');
-  });
-});
