+++ nx-secretarium-website/libs/store/ui-media/README.md
@@ -1,0 +1,7 @@
+# store-ui-media
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test store-ui-media` to execute the unit tests via [Jest](https://jestjs.io).
