+++ nx-secretarium-website/libs/store/ui-media/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  displayName: 'store-media',
+  displayName: 'store-ui-media',
   preset: '../../../jest.preset.js',
   globals: {
     'ts-jest': {
@@ -10,5 +10,5 @@ module.exports = {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/store/media',
+  coverageDirectory: '../../../coverage/libs/store/ui-media',
 };
