+++ nx-secretarium-website/libs/store/ui-media/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/store-ui-media';
