+++ nx-secretarium-website/libs/store/ui-media/src/lib/store-ui-media.spec.ts
@@ -1,0 +1,7 @@
+import { storeUiMedia } from './store-ui-media';
+
+describe('storeUiMedia', () => {
+  it('should work', () => {
+    expect(storeUiMedia()).toEqual('store-ui-media');
+  });
+});
