+++ nx-secretarium-website/tsconfig.base.json
@@ -21,6 +21,9 @@
       "@nx-secretarium-website/store/media": ["libs/store/media/src/index.ts"],
       "@nx-secretarium-website/store/ui-data": [
         "libs/store/ui-data/src/index.ts"
+      ],
+      "@nx-secretarium-website/store/ui-media": [
+        "libs/store/ui-media/src/index.ts"
       ]
     }
   },
